tree: Implement a semi-hacky way to distinguish signed and unsigned
Steal a bit in the tree tag in order to mark if a Natural holds a signed or
unsigned value. This makes it possible for type casts to be a bit smarter about
detecting the correct range.

+++ include/scanner.h
@@ -147,7 +147,7 @@ enum token_t
 
 
 typedef std::vector<uint> indent_list;
-typedef ulong TreePosition;
+typedef uintptr_t TreePosition;
 
 struct Positions
 // ----------------------------------------------------------------------------
